+++ test-packages/e2e-tests/test/openapi.spec.ts
@@ -6,13 +6,12 @@ import { destination } from './test-util';
 
 // TODO: How do I handle paths in rest requests?
 // TODO: Transpilation needed + tsconfig needs dom typings
-// TODO: Response is never "parsed"
 const restDestination = { ...destination, url: destination.url + 'openapi' };
 describe('openapi request builder', () => {
   it('executes getAll request', async () => {
     const request = TestServiceApi.getAllEntities();
     expect(
-      (await request.execute(restDestination)).data.length
+      (await request.execute(restDestination)).length
     ).toBeGreaterThanOrEqual(4);
   });
 
@@ -22,12 +21,12 @@ describe('openapi request builder', () => {
       keyProperty: 'a13e7a92-cb9c-8f4a-b1e1-6b5c8458a9fb',
       stringProperty: 'string'
     };
-    await TestServiceApi.createEntity(testEntity).execute(restDestination);
+    await TestServiceApi.createEntity({ testEntity }).execute(restDestination);
     const postCreateCount = await countEntities();
     expect(postCreateCount).toEqual(preCreateCount + 1);
   });
 });
 
-async function countEntities(): Promise<number> {
-  return (await TestServiceApi.countEntities().execute(restDestination)).data;
+function countEntities(): Promise<number> {
+  return TestServiceApi.countEntities().execute(restDestination);
 }